--- src/presenter/film-list-presenter.js
@@ -1,16 +1,0 @@
-import { render, RenderPosition } from '../utils/render';
-import FilmCountView from './view/film-count-view';
-import FiltersAndStatsView from './view/filters-and-stats-view.js';
-import SortView from './view/sort-view';
-import ProfileRankAndAvatarView from './view/profile-and-rank-avatar-view';
-import FilmsSectionView from './view/films-section-view';
-import FilmCardView from './view/film-card-view';
-import ShowMoreButtonView from './view/show-more-button-view';
-import FilmPopupView from './view/film-popup-view';
-import NoFilmView from './view/no-film-view';
-
-export default class FilmsListPresenter {
-  #filmListContainer = null;
-
-  #filmListComponent = new
-}
